Reorganize cli structure to make it debuggable
