+++ package.json
@@ -26,8 +26,9 @@
     "@hapi/joi": "^15.0.3",
     "@hapi/vision": "^5.5.2",
     "@hapi/wreck": "^15.0.1",
-    "commander": "^2.12.2",
-    "deepmerge": "^2.2.1",
+    "commander": "^5.1.0",
+    "configstore": "^5.0.1",
+    "deepmerge": "^4.2.2",
     "fs-extra": "^5.0.0",
     "node-fetch": "^2.6.0",
     "nunjucks": "^3.2.0",
